fix: circular ref issues
