+++ storybook/webpack.config.js
@@ -48,10 +48,6 @@ module.exports = ({ config }) => {
     }),
   );
 
-  // Checking CI value here
-  // eslint-disable-next-line no-console
-  console.log(`process.env.CI =`, process.env.CI);
-
   if (process.env.CI) {
     config.plugins.push(
       new CircularDependencyPlugin({
